Fix rmdir if dir is not empty

+++ src/net/f4fs/filesystem/P2PFS.java
@@ -411,6 +411,11 @@ public class P2PFS
         if (!(p instanceof MemoryDirectory)) {
             return -ErrorCodes.ENOTDIR();
         }
+        
+        if (!((MemoryDirectory) p).getContents().isEmpty()) {
+            return -ErrorCodes.ENOTEMPTY();
+        }
+        
         p.delete();
         return 0;
     }
